String Extension Changes
